modules/p115sftp/pyproject.toml
[tool.poetry]
name = "p115sftp"
version = "0.0.1"
description = "Python 115 SFTP Server."
authors = ["ChenyangGao <wosiwujm@gmail.com>"]
license = "MIT"
readme = "readme.md"
homepage = "https://github.com/ChenyangGao/p115client/tree/main/modules/p115sftp"
repository = "https://github.com/ChenyangGao/p115client/tree/main/modules/p115sftp"
keywords = ["python", "115", "FTP"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.12"
cachedict = ">=0.0.7"
errno2 = ">=0.0.3"
p115client = ">=0.0.8.2"
python-decotools = ">=0.0.4"
paramiko = "*"
richlog_fs = ">=0.0.3"

[tool.poetry.scripts]
p115sftp = "p115sftp.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "p115sftp"
